etl/requirements.txt
yfinance
pandas
kafka-python
oracledb
requests
